Get both issue ID and pull request URL in `update_redmine_status`
We were only getting the Redmine issue ID, which is necessary to make
the request to update the issue. However, we also want the pull request
URL so this can be included in a field in the issue.

Do this by extracting the pull request URL from the PR JSON and
generating a tab-separated list of issue IDs and PR URLs. We then split
those two values and use them in `update_redmine_status`.