Store cache file in `XDG_DATA_HOME`; Rename cache file
Get rid of the temporary location and name of our cache file. Now use
the XDG data directory and create ourselves a subdirectory under it
where we can store our cache files.

We want a separate cache file for each repository you run Redprine
against. To differentiate the cache files, name them with the owner-repo
combination passed in as a command line argument. Get rid of the slash
in the owner-repo string because this is a filename. For example,
'eagle-jump/engine' turns into 'eagle-jump--engine'.