update_redmine_status(): Print error if Redmine request fails
We want the program to error if the update fails, because that's kind of
the whole point.

Thanks to 'pvandenberk' for the explanation of how to get the HTTP
status code from Curl:

https://superuser.com/questions/272265/getting-curl-to-output-http-status-code/442395#442395

    $ curl -s -o /dev/null -w "%{http_code}" http://www.example.org/
    $ curl -s -o /dev/null -I -w "%{http_code}" http://www.example.org/

The error information isn't really enough (we only print the status
code, not a real error message from Redmine), but I'm letting it slide
for laziness.

Now that we're exiting from the program on a Redmine error, it means the
cache won't get updated with successful Redmine updated PRs. We need a
way to only store successfully updated issue PRs to the cache file at
the end of the run.