update_redmine_statuses(): Take pull request JSON instead of TSV
This function now takes an array of pull requests as JSON instead of as
TSV. Doing so allows us to pass a single pull request JSON string to
`update_redmine_status()` and have it deal with extracting the issue
number and pull request URL.

Doing this because I want to be able to incrementally update the cache
file on successful Redmine update.

The functions that extracted the issue ID and pull request URL have been
renamed to singular versions of themselves now that they only treat a
single pull request hash instead of an array of them.

A new function takes an array of pull request hashes and uses `jq` to
output each one on a separate line, enabling us to easily loop through
each one in `update_redmine_statuses()`.