Make the Redmine issue fields to update customisable
Add a new command line argument that accepts a Bash script string that
will output the JSON wanted when making the request to update the
Redmine issue. This allows the JSON to be customised to any fields the
user wants to update instead of the ones I hard-coded.

The param is a Bash-executable string to allow us to pass the pull
request URL in to it for use in the JSON.