Set `$CACHE_FILE` in `preflight()`
We had initialised the `$CACHE_FILE` variable at the start of the script,
but because at that point `$GITHUB_OWNER_REPO` is empty, the cache file
just ends up getting called ".json".

In order to get the name we want, we need to wait until the
`$GITHUB_OWNER_REPO` variable is initialised by the command line
argument.

Initialise `$CACHE_FILE` in `preflight()` because it makes sense to put
it in that kind of setup function. Move `preflight()` after
`parse_arguments()` to ensure the program variables (including
`$GITHUB_OWNER_REPO`) are properly validated and initialised before
running setup code.